+++ packages/commonjs/test/fixtures/form/defaultIsModuleExports-false-no-default/output.js
@@ -1,0 +1,7 @@
+import * as commonjsHelpers from "_commonjsHelpers.js";
+import { __exports as input } from "\u0000fixtures/form/defaultIsModuleExports-false-no-default/input.js?commonjs-exports"
+
+var named = input.named = 3;
+
+export default input;
+export { input as __moduleExports, named };