+++ packages/commonjs/test/fixtures/function/dynamic-require-targets-no-fallback/_config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  description:
+    'throws when there is no require call configured as fallback for dynamicRequireTargets',
+  pluginOptions: {
+    dynamicRequireTargets: ['fixtures/function/dynamic-require-targets-no-fallback/dep1.js']
+  }
+};