+++ packages/dynamic-import-vars/src/index.js
@@ -52,10 +52,13 @@ function dynamicImportVariables({ include, exclude, warnOnError } = {}) {
             // will turn these into chunks automatically
             ms.prepend(
               `function __variableDynamicImportRuntime${dynamicImportIndex}__(path) {
-   switch (path) {
-  ${paths.map((p) => `   case '${p}': return import('${p}');`).join('\n  ')}
-     default: return Promise.reject(new Error("Unknown variable dynamic import: " + path));
-   }
+  switch (path) {
+${paths.map((p) => `    case '${p}': return import('${p}');`).join('\n')}
+${`    default: return new Promise(function(resolve, reject) {
+      (typeof queueMicrotask === 'function' ? queueMicrotask : setTimeout)(
+        reject.bind(null, new Error("Unknown variable dynamic import: " + path))
+      );
+    })\n`}   }
  }\n\n`
             );
             // call the runtime function instead of doing a dynamic import, the import specifier will