+++ packages/eslint/src/index.ts
@@ -7,10 +7,7 @@ import { CLIEngine } from 'eslint';
 import { RollupEslintOptions } from '../types';
 
 function normalizePath(id: string) {
-  return path
-    .relative(process.cwd(), id)
-    .split(path.sep)
-    .join('/');
+  return path.relative(process.cwd(), id).split(path.sep).join('/');
 }
 
 export default function eslint(options = {} as RollupEslintOptions): Plugin {