+++ packages/node-resolve/test/fixtures/browser-exports-browser-browser.js
@@ -1,0 +1,3 @@
+import b from 'exports-browser-browser';
+
+module.exports = b;