+++ packages/run/src/index.ts
@@ -10,8 +10,10 @@ export default function run(opts: RollupRunOptions = {}): Plugin {
   let proc: ChildProcess;
 
   const args = opts.args || [];
+  const allowRestarts = opts.allowRestarts || false;
   const forkOptions = opts.options || opts;
   delete (forkOptions as RollupRunOptions).args;
+  delete (forkOptions as RollupRunOptions).allowRestarts;
 
   return {
     name: 'run',
@@ -41,6 +43,11 @@ export default function run(opts: RollupRunOptions = {}): Plugin {
     },
 
     writeBundle(outputOptions, bundle) {
+      const forkBundle = (dir: string, entryFileName: string) => {
+        if (proc) proc.kill();
+        proc = fork(path.join(dir, entryFileName), args, forkOptions);
+      };
+
       const dir = outputOptions.dir || path.dirname(outputOptions.file!);
       const entryFileName = Object.keys(bundle).find((fileName) => {
         const chunk = bundle[fileName] as RenderedChunk;
@@ -48,8 +55,23 @@ export default function run(opts: RollupRunOptions = {}): Plugin {
       });
 
       if (entryFileName) {
-        if (proc) proc.kill();
-        proc = fork(path.join(dir, entryFileName), args, forkOptions);
+        forkBundle(dir, entryFileName);
+
+        if (allowRestarts) {
+          process.stdin.resume();
+          process.stdin.setEncoding('utf8');
+
+          process.stdin.on('data', (data) => {
+            const line = data.toString().trim().toLowerCase();
+
+            if (line === 'rs' || line === 'restart' || data.toString().charCodeAt(0) === 11) {
+              forkBundle(dir, entryFileName);
+            } else if (line === 'cls' || line === 'clear' || data.toString().charCodeAt(0) === 12) {
+              // eslint-disable-next-line no-console
+              console.clear();
+            }
+          });
+        }
       } else {
         this.error(`@rollup/plugin-run could not find output chunk`);
       }